requirements.txt
fake_bpy_module==20241009
PyQt5==5.15.11
Requests==2.32.3
